Extract useRouteFormatters and useRouteMetrics hooks
Reduce useSavedRouteData from 109 to 49 lines by extracting focused
sub-hooks for formatters and route metrics computation.
