+++ src/app/features/map/saved/use-saved-route-data.ts
@@ -4,43 +4,22 @@ import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
 import type { WalkRouteSummary } from "../../../data/map";
-import { buildDifficultyLookup } from "../../../data/registries/difficulties";
 import { pickLocalization } from "../../../domain/entities/localization";
 import { formatRelativeTime } from "../../../lib/relative-time";
-import { formatDistance, formatDuration, formatStops } from "../../../units/unit-format";
-import { useUnitPreferences } from "../../../units/unit-preferences-provider";
+import { type RouteFormatters, useRouteFormatters } from "./hooks/use-route-formatters";
+import { type RouteMetrics, useRouteMetrics } from "./hooks/use-route-metrics";
 
-export type SavedRouteData = {
-  readonly routeCopy: ReturnType<typeof pickLocalization>;
-  readonly difficultyLabel: string;
-  readonly updatedLabel: string;
-  readonly distance: ReturnType<typeof formatDistance>;
-  readonly duration: ReturnType<typeof formatDuration>;
-  readonly stops: ReturnType<typeof formatStops>;
-  readonly numberFormatter: Intl.NumberFormat;
-  readonly ratingFormatter: Intl.NumberFormat;
-  readonly difficultyLookup: ReturnType<typeof buildDifficultyLookup>;
-};
+export type SavedRouteData = RouteFormatters &
+  RouteMetrics & {
+    readonly routeCopy: ReturnType<typeof pickLocalization>;
+    readonly difficultyLabel: string;
+    readonly updatedLabel: string;
+  };
 
 export const useSavedRouteData = (route: WalkRouteSummary): SavedRouteData => {
-  const { t, i18n } = useTranslation();
-  const { unitSystem } = useUnitPreferences();
-
-  const numberFormatter = useMemo(() => new Intl.NumberFormat(i18n.language), [i18n.language]);
-  const ratingFormatter = useMemo(
-    () =>
-      new Intl.NumberFormat(i18n.language, {
-        minimumFractionDigits: 1,
-        maximumFractionDigits: 1,
-      }),
-    [i18n.language],
-  );
-  const difficultyLookup = useMemo(() => buildDifficultyLookup(t), [t]);
-
-  const unitOptions = useMemo(
-    () => ({ t, locale: i18n.language, unitSystem }),
-    [i18n.language, t, unitSystem],
-  );
+  const { i18n } = useTranslation();
+  const formatters = useRouteFormatters();
+  const metrics = useRouteMetrics(route);
 
   const routeCopy = useMemo(
     () => pickLocalization(route.localizations, i18n.language),
@@ -48,8 +27,8 @@ export const useSavedRouteData = (route: WalkRouteSummary): SavedRouteData => {
   );
 
   const difficultyLabel = useMemo(
-    () => difficultyLookup.get(route.difficultyId)?.label ?? route.difficultyId,
-    [route.difficultyId, difficultyLookup],
+    () => formatters.difficultyLookup.get(route.difficultyId)?.label ?? route.difficultyId,
+    [route.difficultyId, formatters.difficultyLookup],
   );
 
   const updatedLabel = useMemo(
@@ -57,53 +36,14 @@ export const useSavedRouteData = (route: WalkRouteSummary): SavedRouteData => {
     [route.lastUpdatedAt, i18n.language],
   );
 
-  const distance = useMemo(
-    () => formatDistance(route.distanceMetres, unitOptions),
-    [route.distanceMetres, unitOptions],
-  );
-
-  const duration = useMemo(
-    () =>
-      formatDuration(route.durationSeconds, {
-        ...unitOptions,
-        minimumFractionDigits: 0,
-        maximumFractionDigits: 0,
-      }),
-    [route.durationSeconds, unitOptions],
-  );
-
-  const stops = useMemo(
-    () =>
-      formatStops(route.stopsCount, {
-        ...unitOptions,
-        minimumFractionDigits: 0,
-        maximumFractionDigits: 0,
-      }),
-    [route.stopsCount, unitOptions],
-  );
-
   return useMemo(
     () => ({
+      ...formatters,
+      ...metrics,
       routeCopy,
       difficultyLabel,
       updatedLabel,
-      distance,
-      duration,
-      stops,
-      numberFormatter,
-      ratingFormatter,
-      difficultyLookup,
     }),
-    [
-      routeCopy,
-      difficultyLabel,
-      updatedLabel,
-      distance,
-      duration,
-      stops,
-      numberFormatter,
-      ratingFormatter,
-      difficultyLookup,
-    ],
+    [formatters, metrics, routeCopy, difficultyLabel, updatedLabel],
   );
 };
